# imb-catalog-ui/src/app/shared/mapper/product.model.ts
interface Product {
  productId: string;
  productName: string;
  productOwnerName: string;
  developers: Array<string>;
  scrumMasterName: string;
  startDate: string;
  methodology: string;
}